QuickFit3/LIB: QVariant ambiguity
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@511 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ lib/qfrawdatarecord.cpp
@@ -302,7 +302,7 @@ QVariant QFRawDataRecord::resultsGetAsQVariant(QString evalName, QString resultN
     evaluationResult r=resultsGet(evalName, resultName);
     switch(r.type) {
         case qfrdreBoolean: result=r.bvalue; break;
-        case qfrdreInteger: result=r.ivalue; break;
+        case qfrdreInteger: result=(qlonglong)r.ivalue; break;
         case qfrdreNumberError: result=QPointF(r.dvalue, r.derror); break;
         case qfrdreNumber: result=r.dvalue; break;
         case qfrdreNumberList: {
